Refactor: 상수 모음 디렉토리 생성
