+++ constants/metaInfo.js
@@ -1,20 +1,26 @@
-export const POST_DETAIL = {
+const POST_DETAIL = {
   TITLE: '바닐로그 포스트 상세화면',
   DESCRIPTION: '컨텐츠를 불러오지 못했습니다',
 };
 
-export const HOME = {
+const HOME = {
   TITLE: 'VanilLog 바닐로그 홈페이지 메인화면',
   DESCRIPTION: '일상을 VanilLog와 함께 공유하세요',
 };
 
-export const BLOG_HOME = {
+const BLOG_HOME = {
   TITLE: 'VanilLog 바닐로그 블로그 메인화면',
   DESCRIPTION: '나만의 공간',
+};
+
+const DEFAULT = {
+  TITLE: 'VanilLog',
+  DESCRIPTION: 'VanilLog 포스트',
 };
 
 export const METAINFO = {
   POST_DETAIL,
   HOME,
   BLOG_HOME,
+  DEFAULT,
 };
