+++ lib/dbConnect.js
@@ -1,11 +1,10 @@
+import { ERRORS } from 'constants/errors';
 import mongoose from 'mongoose';
 
 const MONGODB_URI = process.env.MONGODB_URI;
 
 if (!MONGODB_URI) {
-  throw new Error(
-    'Please define the MONGODB_URI environment variable inside .env',
-  );
+  throw new Error(ERRORS.MONGODB_URI_NOT_FOUND.MESSAGE);
 }
 
 let cached = global.mongoose;
