+++ src/app/api/v1/post/[postId]/route.js
@@ -5,7 +5,7 @@ import { JSDOM } from 'jsdom';
 import Post from '@models/Post';
 import User from '@models/User';
 import dbConnect from '@lib/dbConnect';
-import { ERRORS } from '@utils/errors';
+import { ERRORS } from 'constants/errors';
 import { sendErrorResponse } from '@utils/response';
 import { validateObjectId } from '@utils/validateObjectId';
 import { findById } from '@utils/findById';
