+++ src/app/api/v1/posts/search/[[...userId]]/route.js
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server';
 import createError from 'http-errors';
 import dbConnect from '@lib/dbConnect';
 import Post from '@models/Post';
-import { ERRORS } from '@utils/errors';
+import { ERRORS } from 'constants/errors';
 import { sendErrorResponse } from '@utils/response';
 import { validateObjectId } from '@utils/validateObjectId';
