+++ src/app/api/v1/profile/updateImage/[userId]/route.js
@@ -1,6 +1,6 @@
 import dbConnect from '@lib/dbConnect';
 import User from '@models/User';
-import { ERRORS } from '@utils/errors';
+import { ERRORS } from 'constants/errors';
 import { NextResponse } from 'next/server';
 import { sendErrorResponse } from '@utils/response';
 import { getLastPartOfUrl } from '@utils/getLastPartOfUrl';
