+++ utils/validateObjectId.js
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 import createError from 'http-errors';
-import { ERRORS } from '@utils/errors';
+import { ERRORS } from 'constants/errors';
 
 export function validateObjectId(userId) {
   if (!mongoose.Types.ObjectId.isValid(userId)) {
